Added Comment section for plugins.

+++ SLPluginDepot/SLPluginDepotModels/Models/PluginRating.cs
@@ -12,7 +12,7 @@ namespace SLPluginDepotModels.Models
         public string UserId { get; set; }
         public ApplicationUser User { get; set; }
 
-        public int Stars { get; set; }
+        public double Stars { get; set; }
         public string Review { get; set; }
 
         public DateTime RatedAt { get; set; }
